pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
]
